backoffice: separation of concerns
