+++ ui/src/backoffice/authors/components/PrivateNotes.tsx
