+++ ui/src/backoffice/authors/components/ResultItem/ResultItem.less
