+++ ui/src/backoffice/authors/containers/AuthorDetailPageContainer.less
