+++ ui/src/backoffice/common/components/DeleteWorkflow/DeleteWorkflow.less
