+++ ui/src/backoffice/dashboard/containers/DashboardPageContainer.less
