+++ ui/src/backoffice/search/containers/SearchPageContainer.less
